signed for windows atomic

+++ src/athr/ovs_atomic_msvc.h
@@ -545,17 +545,17 @@ static inline bool atomic_load_bool(atomic_bool *x)
 
 static inline unsigned long atomic_load_ul(atomic_ulong *x)
 {
-    volatile unsigned long y;
+    volatile long y;
     atomic_read_explicit(x, &y, memory_order_seq_cst);
-    return y;
+    return (unsigned long)y;
 }
 
 static inline unsigned long atomic_fetch_add_ul(atomic_ulong *x,
                                                 unsigned long arg)
 {
-    volatile unsigned long y;
+    volatile long y;
     atomic_add_explicit(x, arg, &y, memory_order_seq_cst);
-    return y;
+    return (unsigned long)y;
 }
 
 #endif
